Re-implement the way keyboard shortcuts work.
Depend on the latest version of xfconf as it includes some bug fixes we
need for this to work.

Add FrapShortcutsProvider which acts as an abstraction layer over the
/default and /custom storage details. A FrapShortCutsProvider
represents a set of shortcuts, /commands or /xfwm4 for instance. It
automatically clones the default settings when created (if necessary)
and provides two signals: 'shortcut-removed' and 'shortcut-added'.

Rework the conflicts dialog, add a reset button the the keyboard
settings dialog and make the whole settings dialog code a GObject
(XfceKeyboardSettings).

(Old svn revision: 28074)