avoid timing out xfce4-session on startup (bug 5040)
xfce4-settings-helper was trying to be clever to migrate itself from
autostart to the actual session.  it was doing a poor job.  this should
keep the cleverness, but make sure it always connects to the session
manager and makes sure the SM can continue.  it also allows the copy of
x-s-h that was started via the session (and not autostart) to force
itself to start, even if it's already running.  the other instance will
lose its hold on the selection, notice, and quit.

(Old svn revision: 29992)