Don't display wrong type and values for properties "with children".
If there is already an iter for a property, make sure we don't skip it
in all cases. If it's a real property, set the type and value, else we
display the "fake" empty property in the view. This fixes a nasty issue:
if /foo/bar was added before /foo, /foo was shown as empty even if it
had a value. This seemed a random issue because the hash table provides
the keys in different orders: sometimes /foo1/bar, /foo1 and some other
times /foo2, /foo2/bar. While the second case worked, the first one
triggered the issue.

Leave the debug statements for now, this still needs to be tested. We
might remove them once we are sure this works fine.