Rework the detection of the current values
Take only the global server config if we are sure that no CRTC at all is
connected to this output. If there is a disabled CRTC connected to it, take the
values from this one, so if we decide to enable it later, these values will be
valid (e.g., we will know which exact rotations are supported).