Refresh the ui when a change is detected
Listen to events from X and rebuild the XfceRandr view and refresh the dialog
when a "screen changed" notification comes up (e.g. when the user plays with
another xrandr tool).

Unfortunately, no event seems to be emitted when an output is physically
{un,}plugged (it will be detected only when the hardware is reprobbed, with
"xrandr -q" or by closing and reopening xfce4-displays-settings).