Display the screen vendor in the ui
Drop the code that tried to read the ConnectorType property, it is not
supported by any driver (although it is marked as mandatory by the xrandr 1.3
spec). Instead, read the screen's EDID data to find its vendor name. It should
be easier to identify a screen in the ui now…