Rework again the detection of supported rotations
If the output is enabled, no problem (no change).

If the output is disabled and connected to some possible crtcs, take the
intersection of the supported rotations for all these crtcs, so when it is
enabled later, the rotation value is sure to be valid.

Else, take the list of supported rotations advertised by the X server.