Rework the UI
Use a single UI for both randr 1.1 and randr 1.2: hide by default some widgets,
and only make them visible with randr 1.2.

Also, with xrandr 1.2, add an on/off radio button to enable/disable an output,
instead of using the resolution for that, it created too many issues; when the
output is off, grey out all comboboxes.