If the user attempted to disable an output, forcefully reload the view
It is possible that it failed because this was the last active output, and
xfce4-settings-helper prevented that from happening. In that case, the screen
state didn't change, no notification is emitted, so xfce4-display-settings will
think that the output is disabled while it is not. Forcefully reloading the
view should avoid that.