Clarify the properties stored in xfconf
In xfce4-display-settings, instead of storing dubious properties like:
 /Default/Output0                      VGA1
 /Default/Output0/Resolution           1280x1024
, store:
 /Default/VGA1                         ViewSonic Corporation 17"
 /Default/VGA1/Resolution              1280x1024
.
This is more obvious, easier to understand and to edit manually.

Adapt xfce4-settings-helper to this change: instead of parsing each property of
the channel and trying to find a match in the currently connected outputs, get
a hash table of all xfconf properties, and for each currently connected output,
lookup for its (possibly) stored properties in the hash table.

(If you are currently testing the branch, you should reset the channel, as
properties from the old scheme "/Default/Output%d" won't be deleted:
"xfconf-query -c displays -p / -r -R".)