Add a rewrite of the xsettings daemon in the helper.
This version is a lot more flexible since it accepts all
properties in the /Xft, /Net and /Gtk, so the user can
add custom properties as well.

The hardcoded defaults have also been removed in favour of a
complete xsettings.xml file and a xfconf-like. It has also
been optimized to steal values from xfconf on startup to
reduce string copying.

The daemon now also does the right thing with serials on properties,
so listeners can update only the changed properties.

In the process a bunch of leaks have been fixed and comments have
been added to explain the xsettings data format, including proper
error trapping of X11 calls.