Fix for --display being fatal (Bug 11188)
When xfsm saves the session it adds a --diplay option to the
applications so they will correcly show up on the correct monitor
when the session starts up again. Commit
791f02416560f5ac75250a9a67f4e2a3d0ac6f7c removed the
g_option_context_add_group (context, gtk_get_option_group... call
because it calls gtk_init before xfsettingsd has a chance to fork
which can cause issues. This patch makes it so that the --display
option is no longer fatal to xfsettingsd which doesn't require
the option anyway as it will apply its settings to all the displays.

Signed-off-by: Sean Davis <smd.seandavis@gmail.com>