Fix the mirror state machine by visualizing all three states
The three states are:
 0: not cloned
 1: cloned (same x/y, same resolution)
 2: mirrored (same x/y, different resolution)

In the case of 2 I added some precautions to make both overlapping
displays visible, so the user knows what the current state is.

TODO: Make sure the resolution combobox is updated correctly when the
mirror checkbox is clicked (currently it remains unaltered, even if the
resolution would change through enabling mirror mode).