xfsettingsd/displays: support xrandr display scaling
Equivalent to xrandr --scale to scale the entire display area

The scale values are stored in xfconf under /Default/<name>/Scale/{X,Y}
similar to Position.

A value of 1.0 is the default with no scaling. Values below 1 will
"zoom in", above 1 will "zoom out". Eg a 1080p monitor and scaling set
to 2x2 will have the same realestate as a 4k panel.

Bug: https://bugzilla.xfce.org/show_bug.cgi?id=14346

Signed-off-by: Jason Zaman <jason@perfinion.com>
Signed-off-by: Sean Davis <smd.seandavis@gmail.com>