display: Disable auto-enabling profiles by default
As this setting is uninitialized and depends on the Notify property and
is thus disabled in the settings UI we should also treat is as disabled
by default in the backend.
This should also mitigate Bug #15437, especially for users upgrading
from Xfce 4.12. (The actual resolution of this bug is still a todo.)