mouse-settings: Fixes reverse horizontal scrolling
The fix was proposed here:
https://bugzilla.xfce.org/show_bug.cgi?id=9689#c4

So I jumped in and made it part of the code, similar to how it was done
for the vertical scroll.

Why? You can think of the usual scroll as moving the scrollbar in a
specific direction: up, down, left, right. A reverse, or natural scroll
is when you want to move the content in a specific direction, which
usually means going against the direction of a scrollbar. Before this
fix, only the vertical was reversed, making the movement unnatural and
confusing.

Fixes: #17
Signed-off-by: Codruț Constantin Gușoi <codrut.gusoi@gmail.com>