appearance: Drop toolbar style setting (Bug #16665)
This setting was relevant when Gtk2 was the main toolkit and toolbars
still respected this setting (or even existed as a widespread UI
element).
Now that we also dropped exotoolbar this setting should also go.

We keep the default setting (style: icon, size: 3) in case users still
have Gtk2 apps that respect this setting.