appearance: Add option to sync xfwm4 theme
Closes #164, Merge request !43

If the "xfwm4" binary is found in path an option is shown to also set
the xfwm4 theme of the same name when a Gtk theme is set.
In addition the Gtk theme list now contains more information on what
types of support a certain theme contains (Gtk3, Gtk2, Xfwm4,
Xfce4-notifyd).