display: Select next preset when minimal dialog is activated (!22)
Useful when calling the dialog via keyboard shorcut and the dialog is
not visible for some reason, for example, a really weird resolution.
At the very least it makes it more comfortable switching presets by
not have to use the mouse.

By "preset" I mean the hardcoded ones:
- External display
- Mirror
- Extend to the right
- Only laptop

This change does not cycle across user-defined profiles.