color-settings: Fix ColorSettings release
The dialog should not be destroyed, since it is owned by the GtkBuilder,
which is released at the end of `main()`. Also, the wrong connection of
`color_settings_dialog_destroy()`, in case the dialog is embedded in the
setting manager, prevented the correct execution of this callback, the
exit from the main loop, and finally the release of the GtkBuilder. It
could also cause segmentation fault, since the pointers accessed were
not the right ones.

Fixes #211.