color-settings: Fix ColorSettings release
The dialog should not be destroyed, since it is owned by the GtkBuilder,
which is released at the end of `main()`. Also, when the dialog is
embedded in the setting manager, the wrong connection of
`color_settings_dialog_destroy()` causes segmentation fault, since the
pointers accessed are not the right ones.

Fixes #211
MR !65