color-settings: Destroy dialog before releasing its GtkBuilder
This amends 4bc8ef847a1af8218a3cff506f787b41e52009b9, where the dialog
destruction was improperly removed. Although this is of no consequence
in this case, since the dialog is identified with the main program, it
is the user's responsibility to destroy toplevel windows from a
GtkBuilder, according to the documentation and contrary to what was said
in that commit.

Related: !65
MR: !70