display: Check for XError on transform apply
Even with XRandR 1.3, not all DDXen support transformations, such as,
for example Xwayland rootful.

Unfortunately, there is no way in the XRandR protocol for the clients to
actually determine that.

And when a client tries to apply a transformation with a DDX which does
not support transformations, the Xserver will raise an XError which will
be fatal to the X11 client if it does not handle such errors.

Even worse, when that happens, the display settings leaves the XRandR
configuration in a broken state where all monitors are removed and
restarting the display settings cannot restore a usable configuration.

To avoid that, make sure to check for an X11 error when applying an
XRandR transformation and log a warning message if that fails.

Signed-off-by: Olivier Fourdan <fourdan@xfce.org>
(cherry picked from commit 319ebf14e438250e6e229dc21491a6ff38912e74)