xfsettingsd: Add XfceGtkSettingsHelper
Synchronizes the "xsettings" Xfconf channel with GtkSettings properties.
This may be of interest on X11 but it is primarily done so that
properties set in Xfconf continue to apply on Wayland.

This commit only adds synchronization via GSettings, for properties
where it is possible. Other properties (e.g. /Gtk/MenuImages) are
synchronized via gtk-modules in a next commit.

Fixes: libxfce4ui#38