xfsettingsd: displays: Always do some actions on RRScreenChangeNotify
Like reloading the output cache, but also some automatic actions, like
switching to the internal output when the external output is
disconnected (especially as the "/Notify" setting only concerns the
connection of new outputs).

In particular, this allows you to revert correctly to a single-output
configuration (internal or external) after switching to extended
configuration (#502).

Fixes: #502