display-settings: Add extend-left to minimal dialog
Not only does this add a useful preset, it also correctly handles the
case where output 0 is on the right and output 1 on the left, which may
not happen often on X11 but does happen on Wayland.

An additional image is added for this purpose and the others are
slightly retouched, the metadata cleaned up and the files renamed
accordingly.

Closes: #100