display-settings: Do not auto-apply profile when opening minimal dialog
There's no reason to do this, as the auto-application of a profile is
limited to the connection of a new output, as an option in the main
dialog. The minimal dialog should only reflect the current setup when
opened, so if a profile is active it should be selected, as with other
presets.