display-settings: Allow to mirror disabled outputs
So enable disabled outputs and switch them to the correct mode on the
fly. This eliminates that weird insensitivity of the mirror button in
the minimal dialog, or the checkbox in the main dialog, when only one
output is active, making it look as if the outputs can't be mirrored
(for lack of compatible modes).