display-settings: Better update output positioning on user interact
There's still room for improvement, for example to fix the gaps or
overlaps that are created between outputs after rotation or scaling. But
this commit is primarily aimed at completing the previous commit, by
correctly positioning outputs after their activation/deactivation, now
that this no longer triggers automatic application of changes on the
server side.