Revert "display-settings: Include auto-enable profile in combo box"
This reverts commit 8fd93e6561a54c1d0edb956aea67966ce3088965.

Contrary to what the label indicates, profile auto-activation does not
only take place when a display is connected, but also when a display is
disconnected. What's more, profile auto-activation is not necessarily
exclusive of another choice, as I said in this commit's message, it's
just a priority auto-activation attempt.

It therefore deserves to be reconsidered, which is the subject of the
next commit.