display-settings: Make main dialog unique
Just a small change that extends the uniqueness of the minimal dialog to
the main dialog. This was made all the more necessary by the use of the
main dialog in xfsettingsd in 0ad0d1de. As said in 44ab2126, the use of
GtkApplication could be more complete, but it seems sufficient in this
case.

Fixes: #81
Completes: 44ab21265ea5b10fcc33d05c453c00370a73c844
Related: 0ad0d1de54f192f345958c35924e6d55a60744c5