xfsettingsd: Try to match the default display profile as well
If profile matching works correctly (which isn't always the case, but
needs to be fixed/completed separately), this should make it possible to
restore the last known configuration in various situations, such as when
a temporarily connected/disconnected screen is disconnected/reconnected
(#57).

It should also render invisible to the user any unintentional screen
disconnections that may occur when the computer is put to sleep or the
screen is turned off (#32, #420).

The default value of the /AutoEnableProfiles property is set to ALWAYS
to ensure that this behavior is enabled by default wherever possible.

The attempt to restore the last known configuration at start-up remains
unconditional, due to the risk of profile matching failure mentioned
above, but it may be better to leave this particular case out of
/AutoEnableProfiles anyway.

Fixes: #32, #57, #420