display-settings: Reset profile before saving
And always save all outputs, consequently. This completes 9160e07a,
because if an old output remains in the default profile after saving,
the profile will no longer match on the next restore attempt. But this
actually applies to other profiles, so it's possible that this fixes
some bugs at this level, which users were getting around by deleting and
then recreating a profile.