display-settings: Reverse Default <-> Fallback logic to apply changes
This became all the more necessary after 9160e07a, as updating the
default profile after any layout change can interfere with the operation
of xfsettingsd if the display settings dialog is opened in parallel
(with race condition and potentially some pretty ugly stuff, since the
profile is reset before saving).

So now we leave this default profile alone, saving changes only when
they are validated by the user, and for temporary changes we use a
temporary profile.