xfsettingsd: Use default display profile when several profiles match
Normally, in this situation, the user should rather use the dialog to
choose which profile to apply, rather than leave it to automatic
application. But if this is the case and the default profile matches,
applying it seems a reasonable choice, rather than doing nothing.

Completes: 9160e07a12001f8ac066fe7b5c299ca58b0cd9ae