Revert "x11: displays: Disable new outputs if /Notify is set to "Do nothing""
This reverts commit e8657c492fff13d508e9e111fd8dd8b117b49ec9.

There's that xrandr bug again, which disables certain screens when
they're switched off, then re-enables them when they're switched on. So
if you disable them then, they're never re-enabled.

This is solved on master by automatically reapplying the default profile
when a new screen is connected, but unfortunately it will have to stay
that way on 4.18.

Fixes: #535