Enumerate/load icon themes off of the UI thread
On slower media, there can be noticeable pause of even many tens of
seconds on startup as the icon theme preview code loads each icon theme
and renders previews.

Instead, use GIO to do all the file enumeration asynchronously, and then
load the icon themes (this is the *really* slow part) and render the
previews on a separate thread.

Closes #540