x11: displays: Delay RRScreenChangeNotify handling
This is yet another in a series of commits concerning the handling of
the RRScreenChangeNotify event (the latest of which was eb3ca062), which
causes so many problems with configuration resets and/or screens that
refuse to turn off (that come out of sleep as soon as they're put to
sleep). I'm now confronted with this latter problem (perhaps as a result
of a kernel or driver change), and the only way I can see to get around
it without completely disabling any action on connection or
disconnection of a new screen, is to defer processing of the
RRScreenChangeNotify event, so that xrandr's internal machinations on
screen shutdown are transparent to us.

Despite the indentation changes that muck up the diff, this change
simply moves the RRScreenChangeNotify signal handler code into an idle
function, without changing anything else. The next commit does a little
refactoring to clarify the code a bit and make it less error-prone, by
reducing the scope of variables.