displays: Always enable first output
It doesn't seem to make much sense to activate no output, when all have
been disconnected, and one is reconnected. The "do nothing when a
display is connected" option should probably have an exception in this
case in general.

In the case of X11, this remedies yet another instance of xrandr
disconnecting an output when it goes to sleep: when there is only one
output and no default profile has yet been defined (in the case of a new
installation).

Fixes: #567