build: Fix -Wmaybe-uninitialized
This warning appears when building with the default makepkg environment
on arch linux, but I was unable to reproduce this behavior by setting
the flags appearing in makepkg.conf, so I don't know exactly what it's
due to. Anyway, it's a small fix that makes sense, so let's get on with
it.

Fixes: #575