wayland: displays: Don't wl_display_roundtrip() on new output
This can lead to a lot of code being executed before the EDID is
calculated for the new output, in particular the xfsettingsd code (in
zwlr_output_manager_v1_listener.done()) which needs it to match the
display profiles.

So it's better to let the compositor do things at its own pace, and call
the client's XfceWlrOutputListener from a default .done() listener where
you update the data for new outputs that require it (EDID, new status,
etc.), rather than connecting it directly.

The above problem arises in particular with labwc 0.9.0 which, perhaps
because of wlroots 0.19, disconnects then reconnects virtually all
outputs when tty changes, causing xfsettingsd to break the display
setup as it cannot match the default profile (or any other profile).