display-settings: Delete popup dialog asking to update profile
It's confusing; most of the time, you wonder why this dialog shows up.
Even if it were smarter (taking into account the last modified profile
instead of the initial one, and comparing the profile content to the
current setup), it's not obvious that the user would want to be asked
about this when closing the main dialog. It is better to let him
manually update one of the profiles if he wishes.