xfsettingsd: Add option to disable synchronization with GSettings
This allows to use Xfce without modifying GTK settings for other sessions.
If the environment variable XFSETTINGSD_NO_GSETTINGS_SYNC is set, the
GSettings values are left untouched by xfsettingsd.

When synchronization with GSettings is disabled, the settings are still
applied in Xfce via XSettings on X11 and via gtk-module on Wayland.